Clear up contracts and tests

+++ contracts/layerzero/IOFTCoreV2.sol
@@ -16,7 +16,13 @@ interface IOFTCore is IERC165 {
      * _useZro - indicates to use zro to pay L0 fees
      * _adapterParam - flexible bytes array to indicate messaging adapter services in L0
      */
-    function estimateSendFee(uint16 _dstChainId, bytes calldata _toAddress, uint _amount, bool _useZro, bytes calldata _adapterParams) external view returns (uint nativeFee, uint zroFee);
+    function estimateSendFee(
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256 _amount,
+        bool _useZro,
+        bytes calldata _adapterParams
+    ) external view returns (uint256 nativeFee, uint256 zroFee);
 
     /**
      * @dev send `_amount` amount of token to (`_dstChainId`, `_toAddress`) from `_from`
@@ -28,12 +34,20 @@ interface IOFTCore is IERC165 {
      * `_zroPaymentAddress` set to address(0x0) if not paying in ZRO (LayerZero Token)
      * `_adapterParams` is a flexible bytes array to indicate messaging adapter services
      */
-    function sendFrom(address _from, uint16 _dstChainId, bytes calldata _toAddress, uint _amount, address payable _refundAddress, address _zroPaymentAddress, bytes calldata _adapterParams) external payable;
+    function sendFrom(
+        address _from,
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256 _amount,
+        address payable _refundAddress,
+        address _zroPaymentAddress,
+        bytes calldata _adapterParams
+    ) external payable;
 
     /**
      * @dev returns the circulating amount of tokens on current chain
      */
-    function circulatingSupply() external view returns (uint);
+    function circulatingSupply() external view returns (uint256);
 
     /**
      * @dev returns the address of the ERC20 token
@@ -44,13 +58,22 @@ interface IOFTCore is IERC165 {
      * @dev Emitted when `_amount` tokens are moved from the `_sender` to (`_dstChainId`, `_toAddress`)
      * `_nonce` is the outbound nonce
      */
-    event SendToChain(uint16 indexed _dstChainId, address indexed _from, bytes _toAddress, uint _amount);
+    event SendToChain(
+        uint16 indexed _dstChainId,
+        address indexed _from,
+        bytes _toAddress,
+        uint256 _amount
+    );
 
     /**
      * @dev Emitted when `_amount` tokens are received from `_srcChainId` into the `_toAddress` on the local chain.
      * `_nonce` is the inbound nonce.
      */
-    event ReceiveFromChain(uint16 indexed _srcChainId, address indexed _to, uint _amount);
+    event ReceiveFromChain(
+        uint16 indexed _srcChainId,
+        address indexed _to,
+        uint256 _amount
+    );
 
     event SetUseCustomAdapterParams(bool _useCustomAdapterParams);
 }
